domain-watcher: move ListerWatcher construction to NewSharedInformer
Inline newListWatchFromNotify into its only caller, NewSharedInformer.
Move the cross-restart failure test to cache_test.go where the
ListerWatcher wiring lives.
