+++ pkg/virt-handler/cache/cache_test.go
@@ -34,6 +34,7 @@ import (
 	"go.uber.org/mock/gomock"
 
 	"k8s.io/apimachinery/pkg/api/equality"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/tools/cache"
 
@@ -438,6 +439,55 @@ var _ = Describe("Domain informer", func() {
 			err = client.SendDomainEvent(watch.Event{Type: watch.Deleted, Object: domain})
 			Expect(err).ToNot(HaveOccurred())
 			Eventually(func(g Gomega) { verifyObj("default/test", nil, g) }, time.Second, 200*time.Millisecond).Should(Succeed())
+		})
+	})
+})
+
+var _ = Describe("Domain watcher ListerWatcher", func() {
+	Context("consecutive failure across watcher restarts", func() {
+		It("should accumulate failures across Watch() calls via ListerWatcher", func() {
+			origMax := notifyServerMaxConsecutiveFails
+			origHealthy := notifyServerHealthyRunTime
+			defer func() {
+				notifyServerMaxConsecutiveFails = origMax
+				notifyServerHealthyRunTime = origHealthy
+			}()
+			notifyServerMaxConsecutiveFails = 10
+			notifyServerHealthyRunTime = 1 * time.Hour
+
+			failCount := 3
+			consecutiveFails := new(int)
+			runServer := func(_ context.Context, _ chan watch.Event) error {
+				return fmt.Errorf("permanent failure")
+			}
+			lw := &cache.ListWatch{
+				WatchFuncWithContext: func(ctx context.Context, _ metav1.ListOptions) (watch.Interface, error) {
+					return newDomainWatcher(ctx, runServer, 10, 1*time.Hour, nil, consecutiveFails), nil
+				},
+			}
+
+			// Simulate what SharedInformer does: call WatchWithContext(),
+			// drain the result channel, then call it again on failure.
+			// Each call creates a new domainWatcher; the counter
+			// must persist across all of them.
+			ctx := context.Background()
+			for range failCount {
+				w, err := lw.WatchWithContext(ctx, metav1.ListOptions{})
+				Expect(err).ToNot(HaveOccurred())
+				for range w.ResultChan() {
+				}
+			}
+
+			// Retrieve the shared counter from the next watcher.
+			w, err := lw.WatchWithContext(ctx, metav1.ListOptions{})
+			Expect(err).ToNot(HaveOccurred())
+			dw := w.(*domainWatcher)
+			// Wait for this watcher to also finish (it will fail too).
+			for range dw.ResultChan() {
+			}
+			// The counter should reflect all failures, including the
+			// last watcher. If counters are not shared, this will be 1.
+			Expect(*dw.consecutiveFails).To(Equal(failCount + 1))
 		})
 	})
 })
